fixed slider bug, verified all songs params saved

+++ proof_of_concept/apps/iOS/Groover-UI/Groover/GlobalMeasuresCtrl.swift
@@ -56,7 +56,7 @@ class GlobalMeasuresCtrl: SliderCtrl{
     override func draw(_ rect: CGRect) {
         if(self.position == nil){
             self.position = CGPoint(x: 0, y: 0)
-            self.update_pos_from_value(new_value: self.default_value)
+            self.update_pos_from_value(new_value: CGFloat(self.default_value))
             UIGroover.drawSliderCanvas(sliderPosition: self.position.x)
             
         }
